- Something fishy has been detected in the (not so) SSE optimizated   code from SoundTouch's-inspired WSOLA time-stretching. (TESTING)
git-svn-id: http://svn.code.sf.net/p/qtractor/code_attic/trunk@3182 304e7ffa-9037-4501-99c0-529757c9bbcc

+++ src/qtractorTimeStretch.h
@@ -207,7 +207,7 @@ private:
 	bool m_bMidBufferDirty;
 
 	// Calculates the cross-correlation value over the overlap period.
-	double (*m_pfnCrossCorr)(const float *, const float *, unsigned int);
+	float (*m_pfnCrossCorr)(const float *, const float *, unsigned int);
 };
